Added "enable" field to database
Added an "enable" column to the inventory database and added a default test customer to the shop database. Added helper functions to shop.py to allow the setting of a "current_date" for the shop to be set in that will limit the enabled product to things release at and a little before that date.